Refactor domain snap sync
